refactor: proxy state machine decoupled from loop
The proxy state machine used to loop internally, making as much progress
as possible. This made it hard to test some use cases.

Now, the state machine will accept state and return the next
possible state. If an error is returned, we consider it fatal and abort.
If no state is returned, then the state machine can no longer make
progress and the loop will stop.